+++ i18n/SemanticInterlanguageLinks.magic.php
@@ -11,6 +11,7 @@ $magicWords = array();
 $magicWords['en'] = array(
 	'interlanguagelink' => array( 0, 'INTERLANGUAGELINK:', 'SIL:' ),
 	'interlanguagelist' => array( 0, 'INTERLANGUAGELIST:', 'SILIST:' ),
+	'annotatedlanguage' => array( 0, 'ANNOTATEDLANGUAGE:' ),
 );
 
 /**